audio_stream: fix lifetime race in audio_callback + neutrino strict-warning compliance
A race in audio_callback could SIGSEGV when an audio_stream was
destroyed (or moved to a new vector slot via reallocation) while the
SDL audio thread was iterating the mixer's stream snapshot. The
sequence was:

  if (!stream || !stream->m_pimpl) continue;       // null-check passes
  if (!entry.is_valid()) continue;                  // lifetime token passes
  in_use_guard guard = stream->m_pimpl->create_usage_guard();   // crash

Between the lifetime-token check and the m_pimpl deref, the
destructor could free impl out from under us — destructor's
wait-for-inUse=0 only waits for callbacks that already incremented
the guard counter, not ones that just passed the token check and are
about to call create_usage_guard. The token check at line 269 also
flags *destroyed* streams but not *moved-from* ones, so a vector
reallocation in the player's event_streams_ left entry.stream
dangling.

Fix:

  * audio_stream::m_pimpl: unique_ptr<impl> -> shared_ptr<impl>
    so the audio thread can pin the impl alive for the duration
    of an iteration via a parallel weak_ptr held by the mixer.

  * audio_stream::impl::m_lifetime_token: removed. The mixer's
    weak_ptr<void> now aliases m_pimpl directly (via the implicit
    shared_ptr<impl> -> shared_ptr<void> conversion), so storing a
    token *inside* impl would create a self-owning cycle that
    prevents destruction.

  * audio_callback: replaced the racy pattern with
        auto pin = entry.lifetime_token.lock();
        if (!pin) continue;
        auto* stream = static_cast<impl*>(pin.get());
    While `pin` is held, no destructor can free impl. We also stop
    dereferencing entry.stream (audio_stream*) -- that pointer can
    dangle if the host moved its audio_stream into a new container
    slot and the snapshot still holds the old address.

  * Drop user finish/loop callback invocation from audio_callback.
    We no longer have a safe `audio_stream&` to pass them;
    completion is observable via is_playing() polling on the game
    thread.

While in stream.cc:

  * Replace old-style C casts with static_cast in decode_audio,
    run_processors, eval_out_offset, audio_callback, open(),
    restore_state. Make narrowing conversions explicit (size_t ->
    unsigned int, unsigned int -> channels_t, uint64_t ->
    unsigned int).

Also fix neighbouring strict-warning errors so the standalone
build under neutrino's strict warning set is clean again:

  * audio_device.cc: mark unused 'userdata' lambda params.
  * audio_mixer.cc: explicit size_t -> float widening in shrink-
    threshold math; drop pointless inner static_cast<unsigned int>
    that was downcasting size_t round-trip.
  * resampler/resampler_speex.cc: explicit channels_t cast on the
    get_current_channels() result before passing to
    adjust_for_output_spec.

+++ src/musac/audio_mixer.cc
@@ -60,7 +60,8 @@ namespace musac {
             }
 
         } else if (m_allocated_samples > MAX_RETAINED_SAMPLES &&
-                   out_len_samples < static_cast<unsigned int>(m_allocated_samples * SHRINK_THRESHOLD)) {
+                   out_len_samples < static_cast<size_t>(
+                       static_cast<float>(m_allocated_samples) * SHRINK_THRESHOLD)) {
             // Buffer is large and we're using less than 25% of it
             consecutive_small_requests++;
 
@@ -68,7 +69,7 @@ namespace musac {
             if (consecutive_small_requests > STABILITY_FRAMES) {
                 // Calculate new size with headroom for growth
                 size_t new_size = std::max(
-                    static_cast<size_t>(out_len_samples * SHRINK_HEADROOM),
+                    out_len_samples * SHRINK_HEADROOM,
                     MIN_BUFFER_SAMPLES
                 );
 
@@ -79,7 +80,7 @@ namespace musac {
                 m_final_mix_buf.resize(new_size);
                 m_stream_buf.resize(new_size);
                 m_processor_buf.resize(new_size);
-                m_allocated_samples = static_cast<unsigned int>(new_size);
+                m_allocated_samples = new_size;
 
                 // Reset counter
                 consecutive_small_requests = 0;
@@ -88,7 +89,8 @@ namespace musac {
             }
         } else {
             // Reset small request counter if we're using a reasonable amount
-            if (out_len_samples >= static_cast<unsigned int>(m_allocated_samples * SHRINK_THRESHOLD)) {
+            if (out_len_samples >= static_cast<size_t>(
+                    static_cast<float>(m_allocated_samples) * SHRINK_THRESHOLD)) {
                 consecutive_small_requests = 0;
             }
         }
